Resolves #4 - Clean URLs
This resolves Issue #4, which is to use clean URLs (remove the .html
exension) for versus pages. This just makes things friendly for humans
and is less distracting when sharing links with people (shorter is also
nicer). Also, it just wasn't necessary given the rack server config I'm
using anyways, so just fully taking advantage of that.